requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
youtube-transcript-api
sentence-transformers
faiss-cpu
python-dotenv